Backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-dotenv
pymongo
motor
requests
icalendar
